requirements.txt
fastmcp==0.3.0
msal==1.31.1
httpx==0.27.2
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
structlog==24.4.0
click==8.1.8
tenacity==9.0.0
rich==13.9.4
pytest==8.3.4
pytest-asyncio==0.24.0
click
fastmcp
httpx
msal
pydantic
pydantic-settings
pytest
pytest-asyncio
python-dotenv
rich
structlog
tenacity
fastapi
uvicorn
